認可処理を実装

+++ app/Http/Controllers/IngredientsController.php
@@ -3,10 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Models\Ingredients;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class IngredientsController extends Controller
 {
+    use AuthorizesRequests;
+
     public function store(Request $request)
     {
         $validated = $request->validate([
@@ -15,7 +18,7 @@ class IngredientsController extends Controller
         ]);
         $newIngredient = Ingredients::create($validated);
         $updatedIngredients = Ingredients::where('dish_id', $newIngredient->dish_id)->get();
-        
+
         return response()->json($updatedIngredients);
     }
 
@@ -23,6 +26,7 @@ class IngredientsController extends Controller
     {
         $ingredient = Ingredients::find($id);
         $dishId = $ingredient->dish_id;
+        $this->authorize('delete', $ingredient);
         $ingredient->delete();
         $updatedIngredients = Ingredients::where('dish_id', $dishId)->get();
